Add Docker Configuration for Discovery Server
This commit introduces the following changes to the codebase:

- Adds a new `application-docker.properties` file in the `src/main/resources` directory.
- Configures the server port to `8761` using the `server.port` property.
- Sets the Zipkin tracing endpoint to `http://zipkin:9411` using the `management.zipkin.tracing.endpoint` property.

Purpose
The purpose of this commit is to introduce Docker-specific configuration properties for the Discovery Server service. These properties are intended to be used when running the application in a Docker container environment. The configurations include setting the server port and the Zipkin tracing endpoint specific to the Docker environment. This commit ensures that the Discovery Server service can properly run and connect to the Zipkin service when deployed in a Docker-based infrastructure.