Fix Zipkin Tracing Endpoint Configuration
This commit introduces the following changes to the codebase:

- Update the `management.zipkin.tracing.endpoint` property in the `application-docker.properties` file to include the correct endpoint path `/api/v2/spans` for Zipkin.

Purpose:

The purpose of this commit is to fix the configuration for the Zipkin tracing endpoint in the Discovery Server when running in a Docker environment. The previous configuration was missing the `/api/v2/spans` path, which is required by Zipkin to properly receive and process the trace data. By updating this property, the Discovery Server will be able to correctly send tracing data to the Zipkin server, enabling distributed tracing and monitoring of requests across the microservices.